HLAE 2.139.0 (2022-10-04T10:20Z)

+++ AfxHookSource/csgo_CSkyBoxView.h
@@ -2,4 +2,4 @@
 
 bool csgo_CSkyBoxView_Draw_Install(void);
 
-float csgo_CSkyBoxView_GetScale(void);
+int csgo_CSkyBoxView_GetScale(void);
